Show session duration as hours and minutes on separate lines
